demo/enginedemo/pyproject.toml
[tool.setuptools]
packages = ["enginedemo"]

[project]
name = "enginedemo"
version = "0.1.0"
description = 'Crunge Engine Demo'
requires-python = ">=3.10"
authors = [{ name = "Kurtis Fields", email = "kurtisfields@gmail.com" }]

dependencies = [
    "loguru ==0.7.2",
    "PyGLM ==2.7.1",
    "trimesh ==4.0.5",
    "imageio ==2.34.0",
]
